+++ src/api_server/impl/SMFSelectionSubscriptionDataRetrievalApiImpl.cpp
@@ -33,7 +33,6 @@
 
 #include "SMFSelectionSubscriptionDataRetrievalApiImpl.h"
 
-#include "udm_client.hpp"
 #include "logger.hpp"
 #include "udm_config.hpp"
 
@@ -72,7 +71,7 @@ void SMFSelectionSubscriptionDataRetrievalApiImpl::get_smf_sel_data(
 
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udm_app->handle_smf_selection_subscription_data_retrieval(
       supi, response_data, http_code, supported_features_str, plmn_id);
